Move MMA pages, make an authenticated endpoint rather than a webhook
